+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+1.1.0 / 2016-07-10
+==================
+
+* Updated **sync** to return more descriptive data 
+
+
+1.0.20 / 2016-07-04
+==================
+
+* Added **sync** helper for updating existing child nodes 
+
+
 1.0.19 / 2016-07-04
 ==================